requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
mongoengine>=0.29,<1.0
pymongo>=4.9,<5.0
PyJWT>=2.9,<3.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.4,<5.0
drf-yasg>=1.21,<2.0
mongomock>=4.1,<5.0
